+++ app/code/Magento/PaypalUk/Model/Api/Nvp.php
@@ -514,7 +514,7 @@ class Magento_PaypalUk_Model_Api_Nvp extends Magento_Paypal_Model_Api_Nvp
         if ($response['RESULT'] != self::RESPONSE_CODE_APPROVED) {
             $message = $response['RESPMSG'];
             $e = new Exception(sprintf('PayPal gateway errors: %s.', $message));
-            Mage::logException($e);
+            $this->_logger->logException($e);
             Mage::throwException(
                 __('PayPal gateway rejected the request. %1', $message)
             );