+++ app/code/Magento/Pbridge/Model/Payment/Method/Paypaluk.php
@@ -56,6 +56,7 @@ class Magento_Pbridge_Model_Payment_Method_Paypaluk extends Magento_PaypalUk_Mod
     /**
      * Constructor
      *
+     * @param Magento_Core_Model_Logger $logger
      * @param Magento_Pbridge_Helper_Data $pbridgeData
      * @param Magento_Core_Model_Event_Manager $eventManager
      * @param Magento_Core_Model_Store_Config $coreStoreConfig
@@ -64,6 +65,7 @@ class Magento_Pbridge_Model_Payment_Method_Paypaluk extends Magento_PaypalUk_Mod
      * @param array $data
      */
     public function __construct(
+        Magento_Core_Model_Logger $logger,
         Magento_Pbridge_Helper_Data $pbridgeData,
         Magento_Core_Model_Event_Manager $eventManager,
         Magento_Core_Model_Store_Config $coreStoreConfig,
@@ -72,7 +74,7 @@ class Magento_Pbridge_Model_Payment_Method_Paypaluk extends Magento_PaypalUk_Mod
         array $data = array()
     ) {
         $this->_pbridgeData = $pbridgeData;
-        parent::__construct($eventManager, $moduleList, $paymentData, $coreStoreConfig, $data);
+        parent::__construct($logger, $eventManager, $moduleList, $paymentData, $coreStoreConfig, $data);
         $this->_pro->setPaymentMethod($this);
     }